backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sentence-transformers
faiss-cpu
numpy
langgraph
pydantic